@@ -1,12 +1,11 @@
-import '../../app/app.locator.dart';
-import 'package:example/enums/snackbar_type.dart';
 import 'package:flutter/material.dart';
 import 'package:stacked_services/stacked_services.dart';
+import 'package:stacked_services_example/enums/snackbar_type.dart';
 
 import '../../app/app.locator.dart';
 
 class SnackbarView extends StatelessWidget {
-  SnackbarView({Key key}) : super(key: key);
+  SnackbarView({Key? key}) : super(key: key);
 
   final _snackbarService = locator<SnackbarService>();
 
