@@ -1,17 +1,41 @@
+import 'package:example/app/app.locator.dart';
+import 'package:example/app/app.logger.dart';
+import 'package:example/app/app.router.dart';
+import 'package:example/ui/bottom_nav/favorites/favorites_view.dart';
+import 'package:example/ui/bottom_nav/history/history_view.dart';
 import 'package:stacked/stacked.dart';
+import 'package:stacked_services/stacked_services.dart';
 
-class BottomNavExampleViewModel extends BaseViewModel {
-  int _currentTabIndex = 0;
-  int get currentTabIndex => _currentTabIndex;
+class BottomNavExampleViewModel extends IndexTrackingViewModel {
+  final log = getLogger('BottomNavExampleViewModel');
+  final _navigationService = exampleLocator<NavigationService>();
 
-  bool _reverse = false;
-  bool get reverse => _reverse;
-
-  void setTabIndex(int value) {
-    if (value < _currentTabIndex) {
-      _reverse = true;
+  void handleNavigation(int index) {
+    log.i('handleNavigation: $index');
+    setIndex(index);
+    switch (index) {
+      case 0:
+        _navigationService.replaceWithTransition(
+          const FavoritesView(),
+          transitionStyle:
+              reverse ? Transition.rightToLeft : Transition.leftToRight,
+          id: 1,
+        );
+        break;
+      case 1:
+        _navigationService.clearStackAndShowView(
+          const HistoryView(),
+          id: 1,
+        );
+        break;
+      case 2:
+        _navigationService.pushNamedAndRemoveUntil(
+          BottomNavExampleRoutes.profileView,
+          predicate: (route) => route.isFirst,
+          id: 1,
+        );
+        break;
+      default:
     }
-    _currentTabIndex = value;
-    notifyListeners();
   }
 }
