@@ -4,7 +4,7 @@
 // StackedFormGenerator
 // **************************************************************************
 
-// ignore_for_file: public_member_api_docs
+// ignore_for_file: public_member_api_docs,  constant_identifier_names, non_constant_identifier_names,unnecessary_this
 
 import 'package:flutter/material.dart';
 import 'package:stacked/stacked.dart';
@@ -12,11 +12,11 @@ import 'package:stacked/stacked.dart';
 const String CountryValueKey = 'country';
 const String ProvinceValueKey = 'province';
 
-const Map<String, String> CountryValueToTitleMap = {
+final Map<String, String> CountryValueToTitleMap = {
   'ZAR': 'South Africa',
   'UK': 'United Kingdom',
 };
-const Map<String, String> ProvinceValueToTitleMap = {
+final Map<String, String> ProvinceValueToTitleMap = {
   'Western Cape': 'Western Cape',
   'Easter Cape': 'Easter Cape',
   'Free State': 'Free State',
