@@ -1,5 +1,5 @@
+import 'package:example/ui/form/validators.dart';
 import 'package:flutter/material.dart';
-import 'package:new_architecture/ui/form/validators.dart';
 import 'package:stacked/stacked.dart';
 import 'package:stacked/stacked_annotations.dart';
 
@@ -15,7 +15,6 @@ import 'example_form_viewmodel.dart';
   ),
   FormTextField(
     name: 'password',
-    isPassword: true,
     validator: FormValidators.passwordValidator,
   ),
   FormTextField(name: 'shortBio'),
@@ -44,13 +43,17 @@ class ExampleFormView extends StatelessWidget with $ExampleFormView {
       onModelReady: (viewModel) {
         // #3: Listen to text updates by calling listenToFormUpdated(model);
         listenToFormUpdated(viewModel);
+        DoYouLoveFoodValueToTitleMap.addAll({'MaybeDr': 'Maybe'});
         viewModel.setDoYouLoveFood(DoYouLoveFoodValueToTitleMap.keys.first);
       },
+      onDispose: (model) => disposeForm(),
       builder: (context, viewModel, child) => Scaffold(
+        appBar: AppBar(
+          title: const Text('Example Form View'),
+          centerTitle: true,
+        ),
         floatingActionButton: FloatingActionButton(
-          onPressed: () {
-            viewModel.navigateSomewhere();
-          },
+          onPressed: viewModel.navigateToNewView,
         ),
         body: SizedBox(
           width: MediaQuery.of(context).size.width,
@@ -59,13 +62,13 @@ class ExampleFormView extends StatelessWidget with $ExampleFormView {
               mainAxisAlignment: MainAxisAlignment.center,
               children: <Widget>[
                 ConstrainedBox(
-                  constraints: BoxConstraints(
+                  constraints: const BoxConstraints(
                     maxWidth: 300,
                   ),
                   child: TextFormField(
                     //#4: Set email emailController and focus node
                     controller: emailController,
-                    decoration: InputDecoration(hintText: 'email'),
+                    decoration: const InputDecoration(hintText: 'email'),
                     keyboardType: TextInputType.emailAddress,
                     focusNode: emailFocusNode,
                   ),
@@ -73,17 +76,18 @@ class ExampleFormView extends StatelessWidget with $ExampleFormView {
                 if (viewModel.hasEmailValidationMessage)
                   Text(
                     viewModel.emailValidationMessage!,
-                    style: TextStyle(color: Colors.red),
+                    style: const TextStyle(color: Colors.red),
                   ),
-                SizedBox(height: 15),
+                const SizedBox(height: 15),
                 ConstrainedBox(
-                  constraints: BoxConstraints(
+                  constraints: const BoxConstraints(
                     maxWidth: 300,
                   ),
                   child: TextFormField(
                     //#5: Set password passwordController and focus node
+                    key: const ValueKey('passwordField'),
                     controller: passwordController,
-                    decoration: InputDecoration(hintText: 'password'),
+                    decoration: const InputDecoration(hintText: 'password'),
                     keyboardType: TextInputType.visiblePassword,
                     obscureText: true,
                     focusNode: passwordFocusNode,
@@ -93,11 +97,11 @@ class ExampleFormView extends StatelessWidget with $ExampleFormView {
                 if (viewModel.hasPasswordValidationMessage)
                   Text(
                     viewModel.passwordValidationMessage!,
-                    style: TextStyle(color: Colors.red),
+                    style: const TextStyle(color: Colors.red),
                   ),
-                SizedBox(height: 15),
+                const SizedBox(height: 15),
                 ConstrainedBox(
-                  constraints: BoxConstraints(
+                  constraints: const BoxConstraints(
                     maxWidth: 300,
                   ),
                   child: TextField(
@@ -105,13 +109,13 @@ class ExampleFormView extends StatelessWidget with $ExampleFormView {
                     maxLines: null,
                     keyboardType: TextInputType.multiline,
                     controller: shortBioController,
-                    decoration: InputDecoration(
+                    decoration: const InputDecoration(
                       hintText: 'Tell us a bit more about yourself',
                     ),
                     focusNode: shortBioFocusNode,
                   ),
                 ),
-                SizedBox(height: 15),
+                const SizedBox(height: 15),
                 ElevatedButton(
                   onPressed: () => viewModel.selectBirthDate(
                       context: context,
@@ -124,12 +128,14 @@ class ExampleFormView extends StatelessWidget with $ExampleFormView {
                         : 'Select your Date of birth',
                   ),
                 ),
-                SizedBox(height: 15),
+                const SizedBox(height: 15),
                 Row(
+                  mainAxisAlignment: MainAxisAlignment.center,
                   children: [
-                    Text('Do you love food?'),
-                    SizedBox(width: 15),
+                    const Text('Do you love food?'),
+                    const SizedBox(width: 15),
                     DropdownButton<String>(
+                      key: const ValueKey('dropdownField'),
                       value: viewModel.doYouLoveFoodValue,
                       onChanged: (value) {
                         viewModel.setDoYouLoveFood(value!);
@@ -137,6 +143,7 @@ class ExampleFormView extends StatelessWidget with $ExampleFormView {
                       items: DoYouLoveFoodValueToTitleMap.keys
                           .map(
                             (value) => DropdownMenuItem<String>(
+                              key: ValueKey('$value key'),
                               value: value,
                               child: Text(DoYouLoveFoodValueToTitleMap[value]!),
                             ),
@@ -145,11 +152,11 @@ class ExampleFormView extends StatelessWidget with $ExampleFormView {
                     )
                   ],
                 ),
-                SizedBox(height: 15),
+                const SizedBox(height: 15),
                 if (viewModel.showValidationMessage)
                   Text(
                     viewModel.validationMessage!,
-                    style: TextStyle(color: Colors.red),
+                    style: const TextStyle(color: Colors.red),
                   ),
               ],
             ),
