@@ -1,30 +1,27 @@
-import 'package:new_architecture/app/app.locator.dart';
-import 'package:new_architecture/app/app.logger.dart';
-import 'package:new_architecture/app/app.router.dart';
+import 'package:example/app/app.locator.dart';
+import 'package:example/app/app.logger.dart';
+import 'package:example/app/app.router.dart';
 import 'package:stacked/stacked.dart';
 import 'package:stacked_services/stacked_services.dart';
 
 class HomeViewModel extends BaseViewModel {
   final log = getLogger('HomeViewModel');
-  final NavigationService _navigationService = exampleLocator<NavigationService>();
-
-  HomeViewModel() {
-    log.d('created');
-  }
+  final NavigationService _navigationService =
+      exampleLocator<NavigationService>();
 
   String title = 'default';
+  int counter = 0;
 
   void navigate() {
-    _navigationService.navigateTo(Routes.nonReactiveView);
+    _navigationService.navigateToNonReactiveView();
   }
 
   void initialise() {
     log.i('initialise');
-    title = 'initialised';
+    title = 'Initialised';
     notifyListeners();
   }
 
-  int counter = 0;
   void updateTitle() {
     counter++;
     title = '$counter';
