@@ -5,8 +5,8 @@ import '../../app/app.locator.dart';
 import '../../app/app.router.dart';
 
 class NonReactiveViewModel extends BaseViewModel {
-  String title = 'This should not change';
   final _navigationService = exampleLocator<NavigationService>();
+  String title = 'This should not change';
 
   void updateTitle() {
     title += '. This has changed';
@@ -15,5 +15,12 @@ class NonReactiveViewModel extends BaseViewModel {
 
   void navigateToNewView() {
     _navigationService.navigateTo(Routes.streamCounterView);
+  }
+
+  void navigateBackHome() {
+    _navigationService.clearStackAndShow(
+      Routes.homeView,
+      arguments: HomeViewArguments(title: 'Home'),
+    );
   }
 }
