@@ -11,7 +11,8 @@ class RouterClassGenerator extends BaseGenerator with RouteGeneratorHelper {
   RouterClassGenerator(this._rootRouterConfig);
   @override
   String generate() {
-    writeLine("// ignore_for_file: public_member_api_docs");
+    writeLine(
+        "// ignore_for_file: public_member_api_docs, unused_import, non_constant_identifier_names");
     generateImports(
         [..._rootRouterConfig.routes, ..._rootRouterConfig.subRouters]);
     generateAllRoutersIncludingNestedOnes(
@@ -48,7 +49,6 @@ class RouterClassGenerator extends BaseGenerator with RouteGeneratorHelper {
     }
 
     generateArgumentHolders(routes);
-
     routes.where((element) => element.children.isNotEmpty).forEach((element) {
       generateAllRoutersIncludingNestedOnes(
         element.children,
@@ -57,5 +57,7 @@ class RouterClassGenerator extends BaseGenerator with RouteGeneratorHelper {
         generateNavigationHelper,
       );
     });
+
+    generateExtensionForStronglyTypedNavigation(routes);
   }
 }
