@@ -5,21 +5,21 @@ import 'route_config.dart';
 class AdaptiveRouteConfig extends RouteConfig {
   final String? cupertinoNavTitle;
 
-  const AdaptiveRouteConfig({
-    required super.name,
-    required super.pathName,
-    required super.className,
-    super.fullscreenDialog,
-    super.maintainState,
-    super.returnType,
-    super.parameters,
-    super.guards,
-    super.hasWrapper,
-    super.hasConstConstructor,
-    super.children,
-    super.imports,
-    this.cupertinoNavTitle,
-  });
+  const AdaptiveRouteConfig(
+      {required super.name,
+      required super.pathName,
+      required super.className,
+      super.fullscreenDialog,
+      super.maintainState,
+      super.returnType,
+      super.parameters,
+      super.guards,
+      super.hasWrapper,
+      super.hasConstConstructor,
+      super.children,
+      super.imports,
+      this.cupertinoNavTitle,
+      super.isChild});
 
   @override
   Set<String> registerImports() {
@@ -54,6 +54,7 @@ class AdaptiveRouteConfig extends RouteConfig {
     List<RouteConfig>? children,
     Set<String>? imports,
     String? cupertinoNavTitle,
+    bool? isChild,
   }) {
     return AdaptiveRouteConfig(
       name: name ?? this.name,
