@@ -1,7 +1,6 @@
 import 'package:analyzer/dart/element/element.dart';
 import 'package:source_gen/source_gen.dart';
 import 'package:stacked_core/stacked_core.dart';
-
 import 'package:stacked_generator/import_resolver.dart';
 import 'package:stacked_generator/src/generators/router/route_config/route_config_factory.dart';
 import 'package:stacked_generator/utils.dart';
@@ -22,7 +21,8 @@ class RouteConfigResolver {
     this._importResolver,
   );
 
-  Future<RouteConfig> resolve(ConstantReader stackedRoute) async {
+  Future<RouteConfig> resolve(ConstantReader stackedRoute,
+      {bool isChild = false}) async {
     final dartType = stackedRoute.read('page').typeValue;
     throwIf(
       dartType.element is! ClassElement,
@@ -43,7 +43,6 @@ class RouteConfigResolver {
 
     final classElement = dartType.element as ClassElement;
     final className = toDisplayString(dartType);
-
     final import = _importResolver.resolve(classElement);
     if (import != null) imports.add(import);
 
@@ -58,7 +57,27 @@ class RouteConfigResolver {
 
     final returnType = stackedRoute.objectValue.type;
 
-    if (returnType != null && returnType != 'dynamic') {
+    var baseRouteConfig = RouteConfig(
+      hasWrapper: classElement.allSupertypes
+          .map<String>((el) => toDisplayString(el))
+          .contains('StackedRouteWrapper'),
+      returnType: toDisplayString(returnType!),
+      pathName: pathName,
+      name:
+          stackedRoute.peek('name')?.stringValue ?? toLowerCamelCase(className),
+      maintainState: stackedRoute.peek('maintainState')?.boolValue ?? true,
+      imports: imports,
+      guards: extractedGuards ?? [],
+      className: className,
+      fullscreenDialog:
+          stackedRoute.peek('fullscreenDialog')?.boolValue ?? false,
+      isChild: isChild,
+    );
+
+    /// Check if a return type is provided for example [MaterialRoute<int>()]
+    /// and adds the import for that type other wise is will default to dynamic which
+    /// doesn't needs an import
+    if (baseRouteConfig.processedReturnType != 'dynamic') {
       imports.addAll(_importResolver.resolveAll(returnType));
     }
 
@@ -85,7 +104,7 @@ class RouteConfigResolver {
             hasWrapper: classElement.allSupertypes
                 .map<String>((el) => toDisplayString(el))
                 .contains('StackedRouteWrapper'),
-            returnType: toDisplayString(returnType!),
+            returnType: toDisplayString(returnType),
             pathName: pathName,
             name: stackedRoute.peek('name')?.stringValue ??
                 toLowerCamelCase(className),
