@@ -1,16 +1,20 @@
 import 'package:analyzer/dart/constant/value.dart';
+import 'package:source_gen/source_gen.dart';
 import 'package:stacked_generator/import_resolver.dart';
 import 'package:stacked_generator/route_config_resolver.dart';
-import 'package:source_gen/source_gen.dart';
 
 import 'router_config.dart';
 
 class RouterConfigResolver {
   final ImportResolver _importResolver;
 
-  RouterConfigResolver(this._importResolver);
+  const RouterConfigResolver(this._importResolver);
 
-  Future<RouterConfig> resolve(ConstantReader stackedApp) async {
+  Future<RouterConfig?> resolve(ConstantReader stackedApp) async {
+    /// If routes is not provided return null
+    if (!stackedApp.read('routes').isList) return null;
+
+    final routesDartObjects = stackedApp.read('routes').listValue;
     final generateNavigationExt =
         stackedApp.peek('generateNavigationHelperExtension')?.boolValue ??
             false;
@@ -18,35 +22,35 @@ class RouterConfigResolver {
     final routesClassName =
         stackedApp.peek('routesClassName')?.stringValue ?? 'Routes';
 
-    final nestedRoutes = stackedApp.read('routes').listValue;
-
-    var routerConfig = RouterConfig(
+    final routerConfig = RouterConfig(
       routerClassName: 'StackedRouter',
       routesClassName: routesClassName,
       routeNamePrefix: routeNamePrefix,
       generateNavigationHelper: generateNavigationExt,
     );
 
-    final routes = await _resolveRoutes(routerConfig, nestedRoutes);
-    return routerConfig.copyWith(routes: routes);
+    final extractedRoutes =
+        await _resolveRoutes(routerConfig, routesDartObjects);
+
+    return routerConfig.copyWith(routes: extractedRoutes);
   }
 
   Future<List<RouteConfig>> _resolveRoutes(
-    RouterConfig routerConfig,
-    List<DartObject> routesList,
-  ) async {
+      RouterConfig routerConfig, List<DartObject> routesList,
+      {bool isChild = false}) async {
     final List<RouteConfig> allRoutes = [];
 
     for (var routeDartObject in routesList) {
       final routeReader = ConstantReader(routeDartObject);
       var route = await RouteConfigResolver(
               routerConfig.routeNamePrefix, _importResolver)
-          .resolve(routeReader);
+          .resolve(routeReader, isChild: isChild);
 
       final children = routeReader.peek('children')?.listValue;
 
       if (children?.isNotEmpty ?? false) {
-        final childrenRoutes = await _resolveRoutes(routerConfig, children!);
+        final childrenRoutes =
+            await _resolveRoutes(routerConfig, children!, isChild: true);
         route = route.copyWith(children: childrenRoutes);
       }
       allRoutes.add(route);
