@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import 'package:analyzer/dart/element/element.dart';
 import 'package:stacked_core/stacked_core.dart';
 import 'package:stacked_generator/import_resolver.dart';
@@ -9,19 +11,20 @@ import 'generator/router_class_generator.dart';
 
 class StackedRouterGenerator extends GeneratorForAnnotation<StackedApp> {
   @override
-  dynamic generateForAnnotatedElement(
+  FutureOr<String> generateForAnnotatedElement(
     Element element,
     ConstantReader annotation,
     BuildStep buildStep,
   ) async {
-    var libs = await buildStep.resolver.libraries.toList();
-    var importResolver = ImportResolver(libs, element.source?.uri.path ?? '');
+    final libs = await buildStep.resolver.libraries.toList();
+    final importResolver = ImportResolver(libs, element.source?.uri.path ?? '');
 
-    var routerResolver = RouterConfigResolver(importResolver);
-    final routerConfig = await routerResolver.resolve(
-      annotation,
-    );
+    final routerConfig =
+        await RouterConfigResolver(importResolver).resolve(annotation);
 
-    return RouterClassGenerator(routerConfig).generate();
+    if (routerConfig != null)
+      return RouterClassGenerator(routerConfig).generate();
+    else
+      return '';
   }
 }
