@@ -1,5 +1,11 @@
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:stacked_hooks/stacked_hooks.dart';
-
-void main() {}
+void main() {
+  group('StackedHooksTest -', () {
+    group('test -', () {
+      test('just a place holder', () {
+        expect(true, isTrue);
+      });
+    });
+  });
+}
