@@ -1,10 +1,11 @@
-import 'package:example/app/app.locator.dart';
-import 'package:example/app/app.router.dart';
-import 'package:example/ui/setup_bottom_sheet_ui.dart';
-import 'package:example/ui/setup_dialog_ui.dart';
-import 'package:example/ui/setup_snackbar_ui.dart';
+
 import 'package:flutter/material.dart';
 import 'package:stacked_services/stacked_services.dart';
+import 'package:stacked_services_example/app/app.locator.dart';
+import 'package:stacked_services_example/app/app.router.dart';
+import 'package:stacked_services_example/ui/setup_bottom_sheet_ui.dart';
+import 'package:stacked_services_example/ui/setup_dialog_ui.dart';
+import 'package:stacked_services_example/ui/setup_snackbar_ui.dart';
 
 void main() {
   setupLocator();
@@ -35,30 +36,30 @@ class MyApp extends StatelessWidget {
 
 class _LoggingObserver extends NavigatorObserver {
   @override
-  void didPop(Route route, Route previousRoute) {
+  void didPop(Route route, Route? previousRoute) {
     print(
-        'route.name: ${route?.settings?.name}, previousRoute.name: ${previousRoute?.settings?.name}');
+        'route.name: ${route.settings.name}, previousRoute.name: ${previousRoute?.settings.name}');
     super.didPop(route, previousRoute);
   }
 
   @override
-  void didRemove(Route route, Route previousRoute) {
+  void didRemove(Route route, Route? previousRoute) {
     print(
-        'route.name: ${route?.settings?.name}, previousRoute.name: ${previousRoute?.settings?.name}');
+        'route.name: ${route.settings.name}, previousRoute.name: ${previousRoute?.settings.name}');
     super.didRemove(route, previousRoute);
   }
 
   @override
-  void didPush(Route route, Route previousRoute) {
+  void didPush(Route route, Route? previousRoute) {
     print(
-        'route.name: ${route?.settings?.name}, previousRoute.name: ${previousRoute?.settings?.name}');
+        'route.name: ${route.settings.name}, previousRoute.name: ${previousRoute?.settings.name}');
     super.didPush(route, previousRoute);
   }
 
   @override
-  void didReplace({Route newRoute, Route oldRoute}) {
+  void didReplace({Route? newRoute, Route? oldRoute}) {
     print(
-        'newRoute.name: ${newRoute?.settings?.name}, oldRoute.name: ${oldRoute?.settings?.name}');
+        'newRoute.name: ${newRoute?.settings.name}, oldRoute.name: ${oldRoute?.settings.name}');
     super.didReplace(newRoute: newRoute, oldRoute: oldRoute);
   }
 }
