@@ -1,7 +1,8 @@
-import '../../app/app.locator.dart';
-import 'package:example/enums/snackbar_type.dart';
 import 'package:flutter/material.dart';
 import 'package:stacked_services/stacked_services.dart';
+import 'package:stacked_services_example/enums/snackbar_type.dart';
+
+import '../../app/app.locator.dart';
 
 void setupSnackbarUi() {
   final service = locator<SnackbarService>();
