@@ -1,4 +1,4 @@
-// Mocks generated by Mockito 5.1.0 from annotations
+// Mocks generated by Mockito 5.2.0 from annotations
 // in stacked_themes/test/test_setup.dart.
 // Do not manually edit this file.
 
